Move fetchNotes to utils
This function will be reused by wallet send and wallet notes combine commands.
